tests/basic-api-test/requirements.txt
pytest==8.2.2
pyjwt==2.9.0
python-dotenv==1.0.1
